+++ src/Core/Application/Features/ProxyFeature/DeleteProxy/DeleteProxyHandler.cs
@@ -1,4 +1,5 @@
 ﻿using Application.Layers.Persistence.Repository;
+using Domain.Exceptions;
 using MediatR;
 
 namespace Application.Features.ProxyFeature.DeleteProxy;
@@ -12,8 +13,17 @@ public class DeleteProxyHandler: IRequestHandler<DeleteProxyCommand>
         _proxyRepository = proxyRepository;
     }
 
-    public Task Handle(DeleteProxyCommand request, CancellationToken cancellationToken)
+    public async Task Handle(DeleteProxyCommand request, CancellationToken cancellationToken)
     {
-        return _proxyRepository.DeleteProxy(request.ProxyId);
+        var proxy = await _proxyRepository.GetProxy(request.ProxyId);
+
+        if (proxy.UserId != request.UserId)
+        {
+            throw new NoAccessException(
+                $"У пользователя с таким {request.UserId} UserId нет доступа к прокси с таким ${request.ProxyId} ProxyId",
+                "Proxy");
+        }
+        
+        await _proxyRepository.DeleteProxy(request.ProxyId);
     }
 }
