+++ src/Core/Application/Features/ServerFeature/DeleteServer/DeleteServerHandler.cs
@@ -1,4 +1,5 @@
 ﻿using Application.Layers.Persistence.Repository;
+using Domain.Exceptions;
 using MediatR;
 
 namespace Application.Features.ServerFeature.DeleteServer;
@@ -12,8 +13,17 @@ public class DeleteServerHandler: IRequestHandler<DeleteServerCommand>
         _serverRepository = serverRepository;
     }
 
-    public Task Handle(DeleteServerCommand request, CancellationToken cancellationToken)
+    public async Task Handle(DeleteServerCommand request, CancellationToken cancellationToken)
     {
-        return _serverRepository.DeleteServer(request.ServerId);
+        var server = await _serverRepository.GetServer(request.ServerId);
+
+        if (server.UserId != request.UserId)
+        {
+            throw new NoAccessException(
+                $"У пользователя с таким {request.UserId} UserId нет доступа к серверу с таким ${request.ServerId} ServerId",
+                "Server");
+        }
+        
+        await _serverRepository.DeleteServer(request.ServerId);
     }
 }
